test: improve tests for bundle, full coverage
